# refactor: improve `symbolic_evaluate` performance