Before:
```julia
julia> @be (expr, subs) symbolic_evaluate(_...)
Benchmark: 156 samples with 1 evaluation
min    453.542 μs (2318 allocs: 66.359 KiB)
median 501.958 μs (2318 allocs: 66.359 KiB)
mean   605.138 μs (2318 allocs: 66.359 KiB)
max    2.379 ms (2318 allocs: 66.359 KiB)
```

After:
```julia
julia> @be (expr, subs) symbolic_evaluate(_...)
Benchmark: 330 samples with 1 evaluation
min    269.209 μs (1183 allocs: 28.891 KiB)
median 274.625 μs (1183 allocs: 28.891 KiB)
mean   282.744 μs (1183 allocs: 28.891 KiB)
max    588.542 μs (1183 allocs: 28.891 KiB)
```
## Checklist

- [ ] Appropriate tests were added
- [ ] Any code changes were done in a way that does not break public API
- [ ] All documentation related to code changes were updated
- [ ] The new code follows the
  [contributor guidelines](https://github.com/SciML/.github/blob/master/CONTRIBUTING.md), in particular the [SciML Style Guide](https://github.com/SciML/SciMLStyle) and
  [COLPRAC](https://github.com/SciML/COLPRAC).
- [ ] Any new documentation only uses public API
  
## Additional context

Add any other context about the problem here.